Small update

+++ CSGO/CSGO/Source/Features/Miscellaneous/airstuck.cpp
@@ -5,7 +5,7 @@ void CAirstuck::AirStuck( )
 	if ( eng.GetAttack( ) != ( ACTION_DEFAULT | ACTION_PRESS ) )
 	{
 		CGlobalVars gvMaxTickCount = eng.GetGlobalVars( );
-		gvMaxTickCount.uiTickCount = LONG_MAX;
+		gvMaxTickCount.uiTickCount = INT_MAX;
 		eng.SetGlobalVars( gvMaxTickCount );
 	}
 }
